# Replace ahash with foldhash