--- a/Cargo.toml
+++ b/Cargo.toml
@@ -18,7 +18,6 @@ rust-version = "1.86"
 edition = "2024"
 
 [workspace.dependencies]
-ahash = { version = "0.8.0", default-features = false, features = ["std"] }
 anyhow = "1.0.97"
 ascii_table = "4.0.0"
 atomic-take = "1.0.0"
@@ -35,6 +34,7 @@ dhat = { version = "0.3.3" }
 fallible-iterator = "0.3.0"
 fd-lock = "4.0.0"
 flate2 = { version = "1.1.0", features = ["zlib-rs"] }
+foldhash = "0.1.0"
 gimli = "0.31.0"
 hashbrown = "0.15.1"
 hex = "0.4.0"