--- a/libwild/Cargo.toml
+++ b/libwild/Cargo.toml
@@ -8,7 +8,6 @@ rust-version.workspace = true
 edition.workspace = true
 
 [dependencies]
-ahash = { workspace = true }
 anyhow = { workspace = true }
 atomic-take = { workspace = true }
 bitflags = { workspace = true }
@@ -19,6 +18,7 @@ bytesize = { workspace = true }
 crossbeam-queue = { workspace = true }
 crossbeam-utils = { workspace = true }
 flate2 = { workspace = true }
+foldhash = { workspace = true }
 hashbrown = { workspace = true }
 hex = { workspace = true }
 itertools = { workspace = true }