--- a/libwild/src/save_dir.rs
+++ b/libwild/src/save_dir.rs
@@ -1,8 +1,9 @@
 //! Support for saving inputs for later use.
 
 use crate::error::Result;
-use ahash::AHashMap;
 use anyhow::Context;
+use foldhash::HashMap as FoldHashMap;
+use foldhash::HashMapExt as _;
 use std::io::BufWriter;
 use std::io::Write;
 use std::path::Path;
@@ -10,7 +11,7 @@ use std::path::PathBuf;
 
 pub(crate) struct SaveDir {
     dir: Option<PathBuf>,
-    copied_paths: AHashMap<String, String>,
+    copied_paths: FoldHashMap<String, String>,
 }
 
 const SAVE_DIR_ENV: &str = "WILD_SAVE_DIR";
@@ -64,7 +65,7 @@ impl SaveDir {
     fn with_dir(dir: Option<PathBuf>) -> Self {
         SaveDir {
             dir,
-            copied_paths: AHashMap::new(),
+            copied_paths: FoldHashMap::new(),
         }
     }