Hashbrown already did it https://github.com/rust-lang/hashbrown/pull/563

Benchmarks without LTO:
```
~/llvm-tblgen
❯ OUT=/tmp/bin powerprofilesctl launch -p performance poop "./run-with ~/Projects/wild/target/ahash/ld --no-fork" "./run-with ~/Projects/wild/target/foldhash/ld --no-fork"
Benchmark 1 (150 runs): ./run-with ~/Projects/wild/target/ahash/ld --no-fork
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          33.2ms ± 2.91ms    28.7ms … 44.0ms          5 ( 3%)        0%
  peak_rss           46.9MB ± 7.01MB    38.1MB … 71.6MB          4 ( 3%)        0%
  cpu_cycles          994M  ± 68.6M      691M  … 1.16G           8 ( 5%)        0%
  instructions        301M  ± 5.95M      276M  …  316M           6 ( 4%)        0%
  cache_references   11.6M  ±  240K     10.9M  … 12.2M           1 ( 1%)        0%
  cache_misses       3.18M  ± 65.7K     2.98M  … 3.43M           4 ( 3%)        0%
  branch_misses      1.09M  ± 19.6K     1.02M  … 1.14M           8 ( 5%)        0%
Benchmark 2 (153 runs): ./run-with ~/Projects/wild/target/foldhash/ld --no-fork
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          32.8ms ± 2.24ms    29.0ms … 41.4ms          4 ( 3%)          -  1.2% ±  1.8%
  peak_rss           46.1MB ± 5.62MB    38.1MB … 65.1MB          2 ( 1%)          -  1.6% ±  3.1%
  cpu_cycles         1.00G  ± 71.1M      472M  … 1.14G           9 ( 6%)          +  0.8% ±  1.6%
  instructions        301M  ± 8.28M      242M  …  315M           5 ( 3%)          -  0.0% ±  0.5%
  cache_references   11.5M  ±  279K     10.4M  … 12.3M           3 ( 2%)          -  0.5% ±  0.5%
  cache_misses       3.17M  ± 77.7K     2.73M  … 3.36M           3 ( 2%)          -  0.5% ±  0.5%
  branch_misses      1.10M  ± 25.4K      912K  … 1.16M          10 ( 7%)          +  0.4% ±  0.5%

~/clang
❯ OUT=/tmp/bin powerprofilesctl launch -p performance poop "./run-with ~/Projects/wild/target/ahash/ld --no-fork" "./run-with ~/Projects/wild/target/foldhash/ld --no-fork"
Benchmark 1 (36 runs): ./run-with ~/Projects/wild/target/ahash/ld --no-fork
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           139ms ± 3.69ms     135ms …  156ms          1 ( 3%)        0%
  peak_rss            816MB ± 4.72MB     809MB …  830MB          0 ( 0%)        0%
  cpu_cycles         5.88G  ±  135M     5.57G  … 6.17G           0 ( 0%)        0%
  instructions       3.62G  ± 16.8M     3.58G  … 3.65G           0 ( 0%)        0%
  cache_references   82.8M  ±  567K     81.6M  … 83.8M           0 ( 0%)        0%
  cache_misses       23.0M  ±  193K     22.7M  … 23.4M           0 ( 0%)        0%
  branch_misses      8.02M  ± 60.7K     7.94M  … 8.30M           1 ( 3%)        0%
Benchmark 2 (36 runs): ./run-with ~/Projects/wild/target/foldhash/ld --no-fork
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           139ms ± 2.57ms     135ms …  146ms          1 ( 3%)          -  0.0% ±  1.1%
  peak_rss            817MB ± 5.20MB     809MB …  834MB          1 ( 3%)          +  0.0% ±  0.3%
  cpu_cycles         5.96G  ± 95.5M     5.80G  … 6.14G           0 ( 0%)          +  1.3% ±  0.9%
  instructions       3.61G  ± 12.9M     3.59G  … 3.65G           1 ( 3%)          -  0.2% ±  0.2%
  cache_references   83.1M  ±  492K     82.2M  … 84.1M           0 ( 0%)          +  0.4% ±  0.3%
  cache_misses       23.3M  ±  191K     22.9M  … 23.9M           3 ( 8%)          +  1.1% ±  0.4%
  branch_misses      8.07M  ± 28.5K     8.02M  … 8.14M           0 ( 0%)          +  0.6% ±  0.3%
```

With dist profile:
```
❯ OUT=/tmp/bin powerprofilesctl launch -p performance poop "./run-with ~/Projects/wild/target/ahash/ld-dist --no-fork" "./run-with ~/Projects/wild/target/foldhash/ld-dist --no-fork"
Benchmark 1 (154 runs): ./run-with ~/Projects/wild/target/ahash/ld-dist --no-fork
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          32.5ms ± 2.83ms    28.8ms … 42.1ms         11 ( 7%)        0%
  peak_rss           45.4MB ± 6.45MB    38.4MB … 70.8MB          6 ( 4%)        0%
  cpu_cycles          988M  ± 58.3M      786M  … 1.13G           1 ( 1%)        0%
  instructions        329M  ± 10.5M      300M  …  363M           1 ( 1%)        0%
  cache_references   11.6M  ±  274K     10.8M  … 12.4M           3 ( 2%)        0%
  cache_misses       3.16M  ± 73.1K     2.97M  … 3.50M           1 ( 1%)        0%
  branch_misses      1.14M  ± 24.6K     1.06M  … 1.20M           1 ( 1%)        0%
Benchmark 2 (155 runs): ./run-with ~/Projects/wild/target/foldhash/ld-dist --no-fork
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          32.4ms ± 2.89ms    28.5ms … 51.2ms          4 ( 3%)          -  0.3% ±  2.0%
  peak_rss           44.7MB ± 6.17MB    38.3MB … 86.0MB          7 ( 5%)          -  1.6% ±  3.1%
  cpu_cycles          997M  ± 55.7M      725M  … 1.12G           5 ( 3%)          +  0.8% ±  1.3%
  instructions        329M  ± 9.97M      302M  …  365M           2 ( 1%)          +  0.0% ±  0.7%
  cache_references   11.6M  ±  261K     10.9M  … 12.3M           0 ( 0%)          -  0.0% ±  0.5%
  cache_misses       3.18M  ± 62.6K     3.01M  … 3.34M           1 ( 1%)          +  0.7% ±  0.5%
  branch_misses      1.14M  ± 22.4K     1.08M  … 1.20M           3 ( 2%)          +  0.1% ±  0.5%

❯ OUT=/tmp/bin powerprofilesctl launch -p performance poop "./run-with ~/Projects/wild/target/ahash/ld-dist --no-fork" "./run-with ~/Projects/wild/target/foldhash/ld-dist --no-fork"
Benchmark 1 (36 runs): ./run-with ~/Projects/wild/target/ahash/ld-dist --no-fork
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           139ms ± 2.45ms     134ms …  147ms          1 ( 3%)        0%
  peak_rss            817MB ± 6.19MB     810MB …  837MB          5 (14%)        0%
  cpu_cycles         5.92G  ±  109M     5.68G  … 6.15G           0 ( 0%)        0%
  instructions       3.61G  ± 14.0M     3.58G  … 3.64G           0 ( 0%)        0%
  cache_references   83.3M  ±  491K     82.4M  … 84.5M           0 ( 0%)        0%
  cache_misses       23.1M  ±  193K     22.7M  … 23.5M           0 ( 0%)        0%
  branch_misses      8.05M  ± 34.7K     7.99M  … 8.13M           0 ( 0%)        0%
Benchmark 2 (36 runs): ./run-with ~/Projects/wild/target/foldhash/ld-dist --no-fork
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           139ms ± 2.91ms     133ms …  146ms          0 ( 0%)          -  0.1% ±  0.9%
  peak_rss            816MB ± 5.69MB     810MB …  838MB          3 ( 8%)          -  0.2% ±  0.3%
  cpu_cycles         5.94G  ±  115M     5.76G  … 6.15G           0 ( 0%)          +  0.4% ±  0.9%
  instructions       3.61G  ± 13.1M     3.58G  … 3.65G           1 ( 3%)          +  0.0% ±  0.2%
  cache_references   83.1M  ±  470K     82.1M  … 84.2M           1 ( 3%)          -  0.3% ±  0.3%
  cache_misses       23.2M  ±  188K     22.7M  … 23.5M           0 ( 0%)          +  0.4% ±  0.4%
  branch_misses      8.05M  ± 27.5K     8.00M  … 8.11M           0 ( 0%)          -  0.0% ±  0.2%
```

Doesn't seem to make a difference on my PC.